Refactor CLI to use modular command structure
Extracted each subcommand (detect, treesitter, list-ruleset) into
separate modules under whorl/cli/commands/ to simplify the main CLI
and improve code organization.

Changes:
- Created whorl/cli/ directory with __init__.py
- Moved main CLI entry point to whorl/cli/cli.py
- Created whorl/cli/commands/ directory structure
- Moved detect command logic to whorl/cli/commands/detect.py
- Moved treesitter command logic to whorl/cli/commands/treesitter.py
- Moved list-ruleset command logic to whorl/cli/commands/list_ruleset.py
- All tests pass, CLI functionality unchanged

This makes the codebase easier to maintain and extend. The main cli.py
went from ~805 lines to ~58 lines.
