+++ whorl/cli/__init__.py
@@ -1,0 +1,5 @@
+"""CLI module."""
+
+from whorl.cli.cli import main
+
+__all__ = ["main"]
